+++ LineTracer.h
@@ -3,6 +3,7 @@
 #include "PID.h"
 #include "Drive.h"
 #include "Color.h"
+#include <cstdint>
 
 #define RIGHT 1
 #define LEFT -1
@@ -11,11 +12,11 @@ class LineTracer{
 public:
 	LineTracer(Drive *dr, Color *col);
 	int trace(double speed, int edge);
-	void setTarget(int tar);
-	int getTarget();
+	void setTarget(uint8_t tar);
+	uint8_t getTarget();
 
 private:
-	int target;
+	uint8_t target;
 	PID *brightPid;
 	Drive *drive;
 	Color *color;
